requirements.txt
pandas==2.0.0
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0
scikit-learn==1.3.0
geopy==2.4.0 